requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

langchain-core==0.1.28
langchain-text-splitters==0.0.1
langchain-groq==0.0.1
langgraph==0.0.20

pypdf2==3.0.1
pdfplumber==0.10.3
python-docx==1.1.0

pydantic>=2.7
python-dotenv==1.0.0
